* modified Existed validator. Added attribute "boolean empty" with default value - true. If you want to create entity in database with null referenced field, just write @Existed(collection="name_of_specified_collection"). If you want to prohibit creating entity in database with null referenced field, you should write @Existed(collection="name_of_the_specified_collection", empty=false). * modified Template entity. Created constructor Template(Interviewer interviewer). * modified Question entity. maxQuestionValue must have been byte. * modified InterviewQuestion. finalQuestionValue must have been float. * modified Interview. result must have been float. * added Candidate field in Interview entity. * modified tests according to new changes.

+++ src/main/java/com/interview/model/InterviewQuestion.java
@@ -5,20 +5,20 @@ package com.interview.model;
  */
 public class InterviewQuestion extends Question {
 
-    private double finalQuestionValue;
+    private float finalQuestionValue;
 
     private boolean skipped;
 
     public InterviewQuestion() {}
 
-    public InterviewQuestion(Question question, double finalQuestionValue, boolean skipped) {
+    public InterviewQuestion(Question question, float finalQuestionValue, boolean skipped) {
         super(question.getQuestionString(), question.getMaxQuestionValue());
         setId(question.getId());
         setFinalQuestionValue(finalQuestionValue);
         setSkipped(skipped);
     }
 
-    public void setFinalQuestionValue(double value) {
+    public void setFinalQuestionValue(float value) {
         this.finalQuestionValue = value;
     }
